requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2

# Web Application
streamlit==1.29.0
plotly==5.17.0

# SSL/Security
certifi==2023.7.22
